backend/requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
strawberry-graphql==0.289.0
sqlalchemy==2.0.46
asyncpg==0.31.0
alembic==1.11.1
redis==5.0.0
aioredis==2.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.4
structlog==25.5.0
prometheus-client==0.17.0
pydantic==2.12.4
pytest==8.4.2
pytest-asyncio==0.21.0
httpx==0.27.2
httpcore==1.0.6
python-dotenv==1.0.0
requests==2.32.4
tenacity==8.2.2
email-validator==2.0.0
aiosqlite==0.20.0
greenlet==3.2.4
starlette==0.35.1
psycopg2-binary==2.9.11
boto3==1.42.39
pymongo==4.16.0
jsonschema==4.23.0